Changed FastTravelSubPanel definitions to MVC pattern.
